perf(palette): debounce server search, keep previous results while fetching
CommandPalette relied on useDeferredValue as its only rate control, but
useDeferredValue defers re-rendering -- it does not coalesce fetches.
Every keystroke minted a new queryKey and fired a server search request
(typing 14 chars = up to 14 requests against the cold-prone free-tier
backend), and with no placeholderData the Transactions section
flickered empty between keys.

- extract the useDebounce hook (previously local to TransactionFilters)
  to hooks/useDebounce.ts as the single source of truth
- palette: debounce the query 300ms before the queryKey (same pattern
  TransactionFilters already used) + keepPreviousData to hold results
  on screen while the next fetch resolves
